create `getPieChartLegend` function
